Knn model implementation.
The model is pretty inaccurate at the moment and a few things I'd need to test out further down but so far so good it gives a result.

Discovered the inbuilt library options for a lot of the functions so, Scrapped the hard coded version I was trying to do but it still exists as comments if you wanna check it out.

Any help on what I could to improve performance would be hella appreciated. Peace.