Implemented Naive Bayes and cross-validation.
I used sklearn for doing train/test split and cross-validation.
Implemented a naive bayes model with sklearn, and ran 10-fold cross-validation
on it (using tfidf vectors for input). Got mean accuracy of about 91%.

In the course of doing this I found out that sklearn can do basically
everything (tokenization, tfidf, model, train/test split, and
cross-validation) in about ~5 lines of code. This massively reduces the
iteration time for testing models, so we can actually test quite a lot quite
easily if we want.