Naive Bayes Comparison
If you check the first KNN classifer section, scroll down a little bit more, I added a cell that uses my version of train_split_test.. replaced the variables of your own code with mine AND then i compared the results from your version with mine.

The difference in 1 percent which is fairly significant. I'm just wondering why this is the case?