Add Code for Training and Evaluating All Models.
We train a big variety of models to see which suit our data best.
I haven't got results yet because it takes quite a long time to run and
eats up all my RAM.

Because of this, I've committed to a new branch and I'll run it from my desktop
instead of the laptop.