Reword Logistic Regression explanation.
Plus other minor textual changes.
Added note to remember to relate the confusion matrix to the results of
our category similarity section.

Business and [Sport|Politics] are the most similar and thus the hardest
for our model to distinguish between.

(I forgot to add Submission.ipynb to the previous commit)