Cherry-Pick of Lilah's Model Eval Descriptions.
Previously, the commit "Added some descriptions for the model sections"
(hash 49596c0) added valuable descriptions, but managed to undo previous
code history (around TFIDF unit testing addition).

Here, we begin a process of undoing history. We reset the repo to the
state when the TFIDF unit testing existed. Then we cherry pick the
commits in order to reestablish the markdown added after. This will
realign history to the way we want to it to go.

This is the first cherry-picked commit.