Add Section on Making Predictions for Test Set.
Used the final model to make predictions for the test documents and
wrote the results to both a CSV file (content + prediction) and a TXT
file (just the predictions, line by line). Moodle says we need both a
TXT and a CSV so I'm guessing the formats.

Also updated the remove_special_chars() method to remove a-circumflexes,
'â'. These characters appear not due to a fault in the encoding of the
file as we read from it, but actually as raw characters in the file
itself. With respect to this change, the tokens_are_clean() function has
been updated accordingly.